refactor(api: muscle routes): add check before return response and fix catch error for all http requests

+++ app/api/muscle/[...id]/route.js
@@ -1,12 +1,18 @@
 import Muscle from "@modules/server/models/muscle";
-import Exercise from "@modules/server/models/exercise";
 import connectDb from "@lib/mongodb";
 import { NextResponse } from "next/server";
+import checkId from "@modules/server/utils/checkId";
 
 export async function PATCH(request, { params }) {
   try {
     const { id } = params;
+    if (!checkId(id)) {
+      throw { message: "Wrong id", status: 500 };
+    }
     const muscle = await request.json();
+    if (!muscle) {
+      throw { message: "Not found", status: 400 };
+    }
     await connectDb();
     const updatedMuscle = await Muscle.findByIdAndUpdate(
       id,
@@ -14,32 +20,42 @@ export async function PATCH(request, { params }) {
       { new: true },
       { runValidators: true }
     );
-    return NextResponse.json(updatedMuscle, { status: 200 });
+    return NextResponse.json(updatedMuscle, { message: "Muscle updated", status: 200 });
   } catch (err) {
-    const { errors } = err;
-    return NextResponse.json(errors, { status: 404 });
+    const { message, status } = err;
+    return NextResponse.json({ message, status }, { status: status || 404 });
   }
 }
 
 export async function DELETE(request, { params }) {
   try {
     const { id } = params;
+    if (!checkId(id)) {
+      throw { message: "Wrong id", status: 500 };
+    }
     await connectDb();
     const deletedMuscle = await Muscle.findByIdAndDelete(id);
     return NextResponse.json(deletedMuscle);
   } catch (err) {
-    return NextResponse.json({ message: "Error" }, { status: 404 });
+    const { message, status } = err;
+    return NextResponse.json({ message, status }, { status: status || 404 });
   }
 }
 
 export async function GET(request, { params }) {
   try {
     const { id } = params;
+    if (!checkId(id)) {
+      throw { message: "Wrong id", status: 500 };
+    }
     await connectDb();
     const muscle = await Muscle.findById(id);
+    if (!muscle) {
+      throw { message: "Not found", status: 400 };
+    }
     return NextResponse.json(muscle, { status: 200 });
   } catch (err) {
-    const { errors } = err;
-    return NextResponse.json(errors, { status: 404 });
+    const { message, status } = err;
+    return NextResponse.json({ message, status }, { status: status || 404 });
   }
 }
